requirements.txt
flask
flask-cors
pdfplumber
gunicorn
werkzeug
openpyxl
psycopg2-binary
